+++ R/tableby.R
@@ -148,7 +148,15 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
   ## Store this as attribute in the modeldf column, along with the actual name of the variable,
   ## rather than anova(age) showing up in the result (though anova(age) will be the column name in modeldf
   ## but we pull these attributes off later.
-  tmp.fun <- function(x, ...) set_attr(set_attr(x, "name", deparse(substitute(x))), "stats", list(...))
+  tmp.fun <- function(x, ..., digits = NULL, digits.count = NULL, digits.pct = NULL, cat.simplify = NULL, numeric.simplify = NULL)
+  {
+    attr(x, "name") <- deparse(substitute(x))
+    attr(x, "stats") <- list(...)
+    attr(x, "control.list") <- list(digits = digits, digits.count = digits.count, digits.pct = digits.pct,
+                                    cat.simplify = cat.simplify, numeric.simplify = numeric.simplify)
+    x
+  }
+
   for(sp in special)
   {
     if(!is.null(attr(temp.call$formula, "specials")[[sp]])) assign(sp, tmp.fun, envir = tabenv)
@@ -198,6 +206,14 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     by.col <- droplevels(by.col)
     by.levels <- levels(by.col)
   } else by.levels <- sort(unique(by.col))
+  by.col <- as.character(by.col)
+  by.levels <- as.character(by.levels)
+  if(any(by.levels == ""))
+  {
+    warning('Empty string detected in by-variable is not allowed; converting to " ".')
+    by.col[by.col == ""] <- " "
+    by.levels <- unique(replace(by.levels, by.levels == "", " "))
+  }
 
   if(length(by.levels) < 2 && attributes(Terms)$response != 0 && control$test)
   {
@@ -216,6 +232,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
     if(is.null(labelEff))  labelEff <- nameEff
     statList <- list()
     bystatlist <- list()
+    control.list <- attr(currcol, "control.list")
 
     ############################################################
 
@@ -252,7 +269,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
 
     } else if(is.Date(currcol)) {
       ######## Date variable ###############
-      xlevels <- NULL
+      xlevels <- sort(unique(currcol))
 
       ## get stats funs from either formula  or control
       currstats <- control$date.stats
@@ -269,10 +286,10 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
 
     } else if(is.numeric(currcol) || inherits(currcol, "difftime")) {
       ######## Continuous variable (numeric) ###############
-      xlevels <- NULL
 
       ## for difftime, convert to numeric
       if(inherits(currcol, "difftime")) currcol <- as.numeric(currcol)
+      xlevels <- sort(unique(currcol))
 
       ## if no missings, and control says not to show missings,
       ## remove Nmiss stat fun
@@ -295,7 +312,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
       {
         for(bylev in by.levels) {
           idx <- by.col == bylev
-          bystatlist[[as.character(bylev)]] <- do.call(statfun, list(currcol[idx], levels=xlevels, na.rm=TRUE, weights=weights[idx], ...))
+          bystatlist[[bylev]] <- do.call(statfun, list(currcol[idx], levels=xlevels, na.rm=TRUE, weights=weights[idx], ...))
         }
         ## add Total
         bystatlist$Total <- do.call(statfun, list(currcol, levels=xlevels, weights=weights, ...))
@@ -308,7 +325,8 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
       eval(call(currtest, currcol, by.col, chisq.correct=control$chisq.correct, simulate.p.value=control$simulate.p.value, B=control$B))
     } else NULL
 
-    xList[[nameEff]] <- list(stats=statList, test=testout, label=labelEff, name=names(modeldf)[eff], type=vartype)
+    xList[[nameEff]] <- list(stats=statList, test=testout, variable=nameEff, label=labelEff, term=names(modeldf)[eff],
+                             type=vartype, control.list = control.list)
   }
 
   if(length(xList) == 0) stop("No x-variables successfully computed.")
@@ -318,7 +336,7 @@ tableby <- function(formula, data, na.action, subset=NULL, weights=NULL, control
 
   yList <- list()
   yList[[names(modeldf)[1]]] <- list(stats=c(unlist(table(factor(by.col, levels=by.levels), exclude=NA)), Total=sum(!is.na(by.col))),
-                                     label=labelBy, name=names(modeldf)[1])
+                                     label=labelBy, term=names(modeldf)[1])
 
   structure(list(y = yList, x = xList, control = control, Call = match.call(), weights=userWeights), class = "tableby")
 }