+++ tests/testthat/test_compare.R
@@ -20,7 +20,7 @@ df2 <- data.frame(id = paste0("person", 3:1), a = c("c", "b", "a"), b = c(1, 3, 
 
 test_that("Basic comparison works: by row", {
   expect_identical(
-    capture.output(print(compare(df1, df2))),
+    capture.output(compare(df1, df2)),
     c("Compare Object"                                     ,
       ""                                                   ,
       "Function Call: "                                    ,
@@ -38,7 +38,7 @@ test_that("Basic comparison works: by row", {
 
 test_that("Basic comparison works: by id", {
   expect_identical(
-    capture.output(print(compare(df1, df2, by = "id"))),
+    capture.output(compare(df1, df2, by = "id")),
     c("Compare Object"                                     ,
       ""                                                   ,
       "Function Call: "                                    ,
@@ -56,7 +56,7 @@ test_that("Basic comparison works: by id", {
 
 test_that("Basic comparison works: by row.names", {
   expect_identical(
-    capture.output(print(compare(df1, df2, by = "row.names"))),
+    capture.output(compare(df1, df2, by = "row.names")),
     c("Compare Object"                                          ,
       ""                                                        ,
       "Function Call: "                                         ,
@@ -74,7 +74,7 @@ test_that("Basic comparison works: by row.names", {
 
 test_that("Basic comparison works: by row.names for x and something else for y", {
   expect_identical(
-    capture.output(print(compare(df1, df2, by.x = "row.names", by.y = "d"))),
+    capture.output(compare(df1, df2, by.x = "row.names", by.y = "d")),
     c("Compare Object"                                                          ,
       ""                                                                        ,
       "Function Call: "                                                         ,
@@ -94,7 +94,7 @@ df2$listcol <- list(FALSE, "bye now", 1:2)
 
 test_that("List-column comparison works: by id", {
   expect_identical(
-    capture.output(print(compare(df1, df2, by = "id"))),
+    capture.output(compare(df1, df2, by = "id")),
     c("Compare Object"                                     ,
       ""                                                   ,
       "Function Call: "                                    ,
@@ -115,7 +115,7 @@ df2$testdate <- as.Date(c("2017-09-07", "2017-08-08", "2017-09-07"))
 
 test_that("Dates comparison works: by id", {
   expect_identical(
-    capture.output(print(compare(df1, df2, by = "id"))),
+    capture.output(compare(df1, df2, by = "id")),
     c("Compare Object"                                     ,
       ""                                                   ,
       "Function Call: "                                    ,
@@ -133,7 +133,7 @@ test_that("Dates comparison works: by id", {
 
 test_that("Basic mockstudy comparison works: by id", {
   expect_identical(
-    capture.output(print(compare(mockstudy, mockstudy2, by = 'case'))),
+    capture.output(compare(mockstudy, mockstudy2, by = 'case')),
     c("Compare Object"                                                  ,
       ""                                                                ,
       "Function Call: "                                                 ,
@@ -167,7 +167,7 @@ test_that("Using forbidden names throws an error", {
 
 test_that("tol.vars is working correctly", {
   expect_identical(
-    capture.output(print(compare(mockstudy, mockstudy2, by = 'case', tol.vars = "._ "))),
+    capture.output(compare(mockstudy, mockstudy2, by = 'case', tol.vars = "._ ")),
     c("Compare Object"                                                   ,
       ""                                                                 ,
       "Function Call: "                                                  ,
@@ -183,7 +183,7 @@ test_that("tol.vars is working correctly", {
   )
 
   expect_identical(
-    capture.output(print(compare(mockstudy, mockstudy2, by = 'case', tol.vars = c("._ ", "case")))),
+    capture.output(compare(mockstudy, mockstudy2, by = 'case', tol.vars = c("._ ", "case"))),
     c("Compare Object"                                                   ,
       ""                                                                 ,
       "Function Call: "                                                  ,
@@ -206,7 +206,7 @@ tmp <- compare(mockstudy, mockstudy2, by = "case", tol.vars = c("._ ", "case"),
 
 test_that("tolerances are working correctly", {
   expect_identical(
-    capture.output(print(tmp)),
+    capture.output(tmp),
     c("Compare Object"                                                            ,
       ""                                                                          ,
       "Function Call: "                                                           ,
@@ -242,7 +242,7 @@ tmp2 <- compare(mockstudy, mockstudy2, by = "case",
 
 test_that("custom tolerances are working correctly", {
   expect_identical(
-    capture.output(print(tmp2)),
+    capture.output(tmp2),
     c("Compare Object"                                                            ,
       ""                                                                          ,
       "Function Call: "                                                           ,
@@ -265,6 +265,10 @@ test_that("custom tolerances are working correctly", {
 ###########################################################################################################
 
 test_that("helper functions are working correctly", {
+
+  expect_true(n.diffs(compare(mockstudy, mockstudy2, by = "case")) ==
+                n.diffs(summary(compare(mockstudy, mockstudy2, by = "case"))))
+
   expect_true(n.diffs(compare(df1, df2, by = "id")) == n.diffs(summary(compare(df1, df2, by = "id"))))
   expect_identical(diffs(compare(df1, df2, by = "id")), diffs(summary(compare(df1, df2, by = "id"))))
   expect_identical(diffs(compare(df1, df2, by = "id"), by.var = TRUE), diffs(summary(compare(df1, df2, by = "id")), by.var = TRUE))
@@ -281,8 +285,14 @@ test_that("helper functions are working correctly", {
 
 test_that("Summary output looks right (i.e. for factors)", {
   expect_identical(
-    capture.output(summary(compare(mockstudy, mockstudy2, by = "case"))),
-    c(
+    capture.kable(summary(compare(mockstudy, mockstudy2, by = "case"))),
+    c("Table: Summary of data.frames"                                                 ,
+      ""                                                                              ,
+      "version   arg           ncol   nrow"                                           ,
+      "--------  -----------  -----  -----"                                           ,
+      "x         mockstudy       14   1499"                                           ,
+      "y         mockstudy2      13   1495"                                           ,
+      ""                                                                              ,
       ""                                                                              ,
       ""                                                                              ,
       "Table: Variables not shared"                                                   ,
@@ -370,11 +380,92 @@ test_that("Summary output looks right (i.e. for factors)", {
       "race    race    class  "                                                       ,
       "race    race    label  "                                                       ,
       "race    race    levels "                                                       ,
-      "bmi     bmi     label  "                                                       ,
-      ""
+      "bmi     bmi     label  "
     )
   )
 })
+
+
+test_that("Summary output with attributes and max.print options", {
+  expect_identical(
+    capture.kable(summary(compare(mockstudy, mockstudy2, by = "case"), show.attrs = TRUE,
+                          max.print.vars = 2, max.print.obs = 3, max.print.diff = 3, max.print.attrs = 3)),
+    c("Table: Summary of data.frames"                                                ,
+      ""                                                                             ,
+      "version   arg           ncol   nrow"                                          ,
+      "--------  -----------  -----  -----"                                          ,
+      "x         mockstudy       14   1499"                                          ,
+      "y         mockstudy2      13   1495"                                          ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Variables not shared (5 differences not shown)"                        ,
+      ""                                                                             ,
+      "version   variable    position  class     "                                   ,
+      "--------  ---------  ---------  ----------"                                   ,
+      "x         age                2  integer   "                                   ,
+      "x         arm                3  character "                                   ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Other variables not compared"                                          ,
+      ""                                                                             ,
+      "var.x    pos.x  class.x     var.y    pos.y  class.y "                         ,
+      "------  ------  ----------  ------  ------  --------"                         ,
+      "race         5  character   race         3  factor  "                         ,
+      "ast         12  integer     ast          8  numeric "                         ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Observations not shared (1 differences not shown)"                     ,
+      ""                                                                             ,
+      "version     case   observation"                                               ,
+      "--------  ------  ------------"                                               ,
+      "x          88989             9"                                               ,
+      "x          90158             8"                                               ,
+      "x          99508             7"                                               ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Differences detected by variable"                                      ,
+      ""                                                                             ,
+      "var.x         var.y             n   NAs"                                      ,
+      "------------  ------------  -----  ----"                                      ,
+      "sex           sex            1495     0"                                      ,
+      "ps            ps                1     1"                                      ,
+      "hgb           hgb             266   266"                                      ,
+      "bmi           bmi               0     0"                                      ,
+      "alk.phos      alk.phos          0     0"                                      ,
+      "mdquality.s   mdquality.s       0     0"                                      ,
+      "age.ord       age.ord           0     0"                                      ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: First 3 differences detected per variable (1755 differences not shown)",
+      ""                                                                             ,
+      "var.x   var.y     case  values.x   values.y    row.x   row.y"                 ,
+      "------  ------  ------  ---------  ---------  ------  ------"                 ,
+      "sex     sex      76170  Male       Male           26      20"                 ,
+      "sex     sex      76240  Male       Male           27      21"                 ,
+      "sex     sex      76431  Female     Female         28      22"                 ,
+      "ps      ps       86205  0          NA              6       3"                 ,
+      "hgb     hgb      88714  NA         -9            192     186"                 ,
+      "hgb     hgb      88955  NA         -9            204     198"                 ,
+      "hgb     hgb      89549  NA         -9            229     223"                 ,
+      ""                                                                             ,
+      ""                                                                             ,
+      ""                                                                             ,
+      "Table: Non-identical attributes (3 differences not shown)"                    ,
+      ""                                                                             ,
+      "var.x   var.y   name     attr.x                attr.y              "          ,
+      "------  ------  -------  --------------------  --------------------"          ,
+      "sex     sex     label    NA                    Sex (M/F)           "          ,
+      "sex     sex     levels   c(\"Male\", \"Female\")   c(\"Female\", \"Male\") "  ,
+      "race    race    class    NA                    factor              "
+    )
+  )
+})
+